Use-case-Mapping/requirements.txt
openai>=1.3.0
anthropic>=0.8.0
transformers>=4.35.0
torch>=2.0.0
datasets>=2.14.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.66.0
pandas>=2.1.0
numpy>=1.24.0
tiktoken>=0.5.0
colorama>=0.4.6
tabulate>=0.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
